inval fixes
